added an id check for cards

+++ routes/card.py
@@ -9,19 +9,25 @@ def get_all_cards():
 @app.route("/cards/<int:card_id>", methods=["GET"])
 def get_card(card_id):
     card = Card.query.get(card_id)
-    return jsonify(card=card.text, white=card.white)
+    if card:
+        return jsonify(card=card.text, white=card.white)
+    else:
+        return jsonify("No card found with that id.")
 
 @app.route("/cards/<int:card_id>", methods=["PUT"])
 def update_card(card_id):
     card = Card.query.get(card_id)
-    content = request.json
-    if "text" in content:
-        card.text = content["text"]
-    if "white" in content:
-        card.white = content["white"]
-    db.session.add(card)
-    db.session.commit()
-    return jsonify(status="success")
+    if card:
+        content = request.json
+        if "text" in content:
+            card.text = content["text"]
+        if "white" in content:
+            card.white = content["white"]
+        db.session.add(card)
+        db.session.commit()
+        return jsonify(status="success")
+    else:
+        return jsonify("No card found with that id.")
 
 @app.route("/cards/<int:card_id>", methods=["DELETE"])
 def delete_card(card_id):
@@ -30,11 +36,13 @@ def delete_card(card_id):
     deleting a card cascades to all the matches
     it was used in.
     """
-
     card = Card.query.get(card_id)
-    db.session.delete(card)
-    db.session.commit()
-    return jsonify(status="success")
+    if card:
+        db.session.delete(card)
+        db.session.commit()
+        return jsonify(status="success")
+    else:
+        return jsonify("No card found with that id.")
 
 @app.route("/cards", methods=["POST"])
 def create_card():
